hospitals-src/pyproject.toml
[tool.poetry]
name = "poetry init"
version = "0.1.0"
description = ""
authors = ["Steven Kyalo <pykyalo@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2.13"
django-filter = "^24.2"
djangorestframework-gis = "^1.0"
flake8 = "^7.0.0"
black = "^24.4.2"
psycopg2-binary = "^2.9.9"
django-environ = "^0.11.2"
django-extensions = "^3.2.3"
django-leaflet = "^0.30.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
